experience/pyproject.toml
[project]
name = "don-theory-experience"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "japanize-matplotlib>=1.1.3",
    "pandas>=2.1",
]

[tool.uv]
dev-dependencies = []

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
